+++ werkbank/Properties/Resources.Designer.cs
@@ -153,6 +153,16 @@ namespace werkbank.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap btn_reset {
+            get {
+                object obj = ResourceManager.GetObject("btn_reset", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap btn_settings {
             get {
                 object obj = ResourceManager.GetObject("btn_settings", resourceCulture);